implemented the way we start our app
