+++ App.js
@@ -1,40 +1,33 @@
 import React from 'react';
 import { Platform, StatusBar, StyleSheet, View, Text, Button } from 'react-native';
 import { AppLoading, Asset, Font, Icon } from 'expo';
+import AppNavigator from './navigation/AppNavigator';
+// import { createStackNavigator } from 'react-navigation';
+import Header from './components/Header';
+import ContactScreen from './screens/ContactScreen';
+import CreateTaskScreen from './screens/CreateTaskScreen';
+import HomeScreen from './screens/HomeScreen';
+import LoginScreen from './screens/LoginScreen';
+import SignUpScreen from './screens/SignUpScreen';
+import SingleTaskScreen from './screens/SingleTaskScreen';
+import TasksScreen from './screens/TasksScreen';
+import UserProfileScreen from './screens/UserProfileScreen';
+import WelcomeScreen from './screens/WelcomeScreen';
+import CommentScreen from './components/Comment';
+import MapScreen from './screens/MapScreen';
 import { ACTION_MANAGE_DEFAULT_APPS_SETTINGS } from 'expo/build/IntentLauncherAndroid';
 
-// layouts
-import Header from './components/Header';
+// from Jia
+import UploadPhoto from './screens/UploadPhoto';
+import MediaGPS from './screens/MediaGPS';
+import DropMarker from './screens/DropMarker';
+import TasksMapView from './screens/TasksMapView';
+
+
 import TabBarIcon from './components/TabBarIcon';
 import { createStackNavigator, createBottomTabNavigator } from 'react-navigation';
 
-//Tasks Feed
-import TasksScreen from './screens/TasksScreen';
-import SingleTaskScreen from './screens/SingleTaskScreen';
 
-//Tasks Map view
-import TasksMapView from './screens/TasksMapView';
-
-//Create new Task
-import UploadPhoto from './screens/UploadPhoto';
-import MediaGPS from './screens/MediaGPS';
-import CreateTaskScreen from './screens/CreateTaskScreen';
-
-//User Profile
-import UserProfileScreen from './screens/UserProfileScreen';
-
-//Welcome screen
-import HomeScreen from './screens/HomeScreen';
-import WelcomeScreen from './screens/WelcomeScreen';
-import WelcomeScreenT from './screens/WelcomeScreenT';
-
-//User Auth
-import LoginScreen from './screens/LoginScreen';
-import SignUpScreen from './screens/SignUpScreen';
-
-//Contact
-import ContactScreen from './screens/ContactScreen';
-//-- About us
 
 class App extends React.Component {
   state = {
@@ -53,10 +46,10 @@ class App extends React.Component {
     } else {
       return (
         <View style={styles.container}>
-          <Text> Hello </Text>
+          {/* <Text> Hello </Text> */}
     
-          {/* <StatusBar barStyle="default" />
-          <AppNavigator /> */}
+          <StatusBar barStyle="default" />
+          <AppNavigator />
 
         </View>
       );
@@ -97,132 +90,188 @@ const styles = StyleSheet.create({
   },
 });
 
+// const Screens = createStackNavigator({
+//   Home: {
+//    // manipulate the start screen over here
+//     // screen: App, 
+//     screen: App,
+//     navigationOptions: ({ navigation }) => {
+//       title: 'App Test Home'
+//     }
+//   },
+//   CreateTask: {
+//     screen: CreateTaskScreen,
+//     navigationOptions: ({ navigation }) => {
+//       title: 'Create Task'
+//     }
+//   },
+//   LoginScreen: {
+//     screen: LoginScreen,
+//     navigationOptions: ({ navigation }) => {
+//       title: 'Login Screen'
+//     }
+//   },
+//   Comments: {
+//     screen: CommentScreen,
+//     navigationOptions: ({ navigation }) => {
+//       title: 'Comments'
+//     }
+//   },
+//   Contact: {
+//     screen: ContactScreen,
+//     navigationOptions: ({ navigation }) => {
+//       title: 'Contact'
+//     }
+//   },
+//   HomeScreen: {
+//     screen: HomeScreen,
+//     navigationOptions: ({ navigation }) => {
+//       title: 'Home Screen'
+//     }
+//   },
+//   SignUpScreen: {
+//     screen: SignUpScreen,
+//     navigationOptions: ({ navigation }) => {
+//       title: 'Sign Up Screen'
+//     }
+//   },
+//   SingleTaskScreen: {
+//     screen: SingleTaskScreen,
+//     navigationOptions: ({ navigation }) => {
+//       title: 'Single Task Screen'
+//     }
+//   },
+//   // TasksScreen: {
+//   //   screen: TasksScreen,
+//   //   navigationOptions: ({ navigation }) => {
+//   //     title: 'Tasks Screen'
+//   //   }
+//   // },
+//   WelcomeScreen: {
+//     screen: WelcomeScreen,
+//     navigationOptions: ({ navigation }) => {
+//       title: 'Welcome Screen'
+//     }
+//   },
+//   // UserProfileScreen: {
+//   //   screen: UserProfileScreen,
+//   //   navigationOptions: ({ navigation }) => {
+//   //     title: 'User Profile Screen'
+//   //   }
+//   // },
+//   // MapScreen: {
+//   //   screen: MapScreen,
+//   //   navigationOptions: ({ navigation }) => {
+//   //     title: 'Map Screen'
+//   //     }
+//   //   },
+//   // // from Jia
+//   // UploadPhoto: {
+//   //   screen: UploadPhoto,
+//   //   navigationOptions: ({ navigation }) => {
+//   //     title: 'Uplode Media'
+//   //   }
+//   // },
+//   // MediaGPS: {
+//   //   screen: MediaGPS,
+//   //   navigationOptions: ({ navigation }) => {
+//   //     title: 'Media GPS'
+//   //   }
+//   // },
+//   // DropMarker: {
+//   //   screen: DropMarker,
+//   //   navigationOptions: ({ navigation }) => {
+//   //     title: 'Drop Marker'
+//   //   }
+//   // }
+// })
 
-
-//Feed view (Home)
-const WelcomeStack = createStackNavigator({
-  Home:{
-    // change the 
-    screen: WelcomeScreen,
-  },
-  LoginScreen:{
-    screen: LoginScreen,
-  },
-  SignUpScreen:{
-    screen: SignUpScreen,
-  },
-  // App:{
-  //   screen: App,
-  // },
-
-});
-
-WelcomeStack.navigationOptions = {
-  // tabBarVisible: false,
-};
-
-
+// export default Screens;
 
 // Feed view (TasksScreen)
-const FeedStack = createStackNavigator({
-  TasksScreen:{
-    screen: TasksScreen,
-  },
-  // view single task and make comment
-  SingleTaskScreen:{
-    screen: SingleTaskScreen,
-  }, 
-  // log in to Comment (navigate to welcome screen)
-});
+// const FeedStack = createStackNavigator({
+//   TasksScreen:{
+//     screen: TasksScreen,
+//   } 
+// });
 
-FeedStack.navigationOptions = {
-  tabBarLabel: 'Task Feed',
-  tabBarIcon: ({ focused }) => (
-    <TabBarIcon
-      focused={focused}
-      name={Platform.OS === 'ios' ? 'ios-home' : 'md-home'}
-    />
-  ),
-};
+// FeedStack.navigationOptions = {
+//   tabBarLabel: 'Task Feed',
+//   tabBarIcon: ({ focused }) => (
+//     <TabBarIcon
+//       focused={focused}
+//       name={Platform.OS === 'ios' ? 'ios-home' : 'md-home'}
+//     />
+//   ),
+// };
 
 
-// Map View (TasksMapView)
-const MapStack = createStackNavigator({
-  TasksMapView:{
-    screen: TasksMapView,
-  },
-  SingleTaskScreen:{
-    screen: SingleTaskScreen,
-  }, 
-});
+// // Map View (TasksMapView)
+// const MapStack = createStackNavigator({
+//   TasksMapView:{
+//     screen: TasksMapView,
+//   } 
+// });
 
-MapStack.navigationOptions = {
-  tabBarLabel: 'Map View',
-  tabBarIcon: ({ focused }) => (
-    <TabBarIcon
-      focused={focused}
-      name={Platform.OS === 'ios' ? 'ios-pin' : 'md-pin'}
-    />
-  ),
-};
+// MapStack.navigationOptions = {
+//   tabBarLabel: 'Map View',
+//   tabBarIcon: ({ focused }) => (
+//     <TabBarIcon
+//       focused={focused}
+//       name={Platform.OS === 'ios' ? 'ios-pin' : 'md-pin'}
+//     />
+//   ),
+// };
 
 
-// New Task (UploadPhoto)
-const NewTaskStack = createStackNavigator({
-  // upload photo first
-  UploadPhoto:{
-    screen: UploadPhoto,
-  },
-  // then confirm or drop location
-  MediaGPS:{
-    screen: MediaGPS,
-  }, 
-  // fill out the form 
-  CreateTaskScreen:{
-    screen: CreateTaskScreen,
-  } 
+// // New Task (UploadPhoto)
+// const NewTaskStack = createStackNavigator({
+//   UploadPhoto:{
+//     screen: UploadPhoto,
+//   },
+//   MediaGPS:{
+//     screen: MediaGPS,
+//   }, 
+//   // MediaGPS:{
+//   //   screen: MediaGPS,
+//   // } 
   
-});
+// });
 
-NewTaskStack.navigationOptions = {
-  tabBarLabel: 'Create Task',
-  tabBarVisible: false,
-  tabBarIcon: ({ focused }) => (
-    <TabBarIcon
-      focused={focused}
-      name={Platform.OS === 'ios' ? 'ios-add-circle' : 'md-add-circle'}
-    />
-  ),
-};
+// NewTaskStack.navigationOptions = {
+//   tabBarLabel: 'Create Task',
+//   tabBarVisible: false,
+//   tabBarIcon: ({ focused }) => (
+//     <TabBarIcon
+//       focused={focused}
+//       name={Platform.OS === 'ios' ? 'ios-add-circle' : 'md-add-circle'}
+//     />
+//   ),
+// };
 
-// User Profile (UserProfileScreen)
-const UserStack = createStackNavigator({
-  UserProfileScreen:{
-    screen: UserProfileScreen,
-  } 
-});
+// // User Profile (UserProfileScreen)
+// const UserStack = createStackNavigator({
+//   UserProfileScreen:{
+//     screen: UserProfileScreen,
+//   } 
+// });
 
-UserStack.navigationOptions = {
-  tabBarLabel: 'User',
-  tabBarIcon: ({ focused }) => (
-    <TabBarIcon
-      focused={focused}
-      name={Platform.OS === 'ios' ? 'ios-person' : 'md-person'}
-    />
-  ),
-};
+// UserStack.navigationOptions = {
+//   tabBarLabel: 'User',
+//   tabBarIcon: ({ focused }) => (
+//     <TabBarIcon
+//       focused={focused}
+//       name={Platform.OS === 'ios' ? 'ios-person' : 'md-person'}
+//     />
+//   ),
+// };
 
-export default 
+// export default createBottomTabNavigator({
+//   FeedStack,
+//   MapStack,
+//   NewTaskStack,
+//   // WatchStack,
+//   UserStack
+// });
 
-WelcomeStack
-FeedStack
-
-createBottomTabNavigator(
-  {
-  FeedStack,
-  MapStack,
-  NewTaskStack,
-  // WatchStack,
-  UserStack
-},
-);
+export default App;
