+++ navigation/AppNavigator.js
@@ -2,9 +2,15 @@ import React from 'react';
 import { createSwitchNavigator } from 'react-navigation';
 
 import MainTabNavigator from './MainTabNavigator';
+import AuthStack from './AuthNavigator';
 
 export default createSwitchNavigator({
   // You could add another route here for authentication.
   // Read more at https://reactnavigation.org/docs/en/auth-flow.html
+  Auth: AuthStack,
   Main: MainTabNavigator,
-});
+},
+{
+  initialRouteName: 'Auth',
+}
+);
