+++ navigation/MainTabNavigator.js
@@ -3,48 +3,119 @@ import { Platform } from 'react-native';
 import { createStackNavigator, createBottomTabNavigator } from 'react-navigation';
 
 import TabBarIcon from '../components/TabBarIcon';
-import HomeScreen from '../screens/HomeScreen';
-import UserProfileScreen from '../screens/UserProfileScreen';
+//Tasks Feed
+import TasksScreen from '../screens/TasksScreen';
+import SingleTaskScreen from '../screens/SingleTaskScreen';
+
+//Tasks Map view
+import TasksMapView from '../screens/TasksMapView';
+
+//Create new Task
 import UploadPhoto from '../screens/UploadPhoto';
+import MediaGPS from '../screens/MediaGPS';
+import CreateTaskScreen from '../screens/CreateTaskScreen';
 
-const HomeStack = createStackNavigator({
-  Home: UploadPhoto,
-  UploadPhoto: {
+//User Profile
+import UserProfileScreen from '../screens/UserProfileScreen';
+
+
+
+// Feed view (TasksScreen)
+const FeedStack = createStackNavigator({
+  Home:{
+    screen: TasksScreen,
+  },
+  // view single task and make comment
+  SingleTaskScreen:{
+    screen: SingleTaskScreen,
+  }, 
+  // log in to Comment (navigate to welcome screen)
+});
+
+FeedStack.navigationOptions = {
+  tabBarLabel: 'Task Feed',
+  tabBarIcon: ({ focused }) => (
+    <TabBarIcon
+      focused={focused}
+      name={Platform.OS === 'ios' ? 'ios-home' : 'md-home'}
+    />
+  ),
+};
+
+
+// Map View (TasksMapView)
+const MapStack = createStackNavigator({
+  TasksMapView:{
+    screen: TasksMapView,
+  },
+  SingleTaskScreen:{
+    screen: SingleTaskScreen,
+  }, 
+});
+
+MapStack.navigationOptions = {
+  tabBarLabel: 'Map View',
+  tabBarIcon: ({ focused }) => (
+    <TabBarIcon
+      focused={focused}
+      name={Platform.OS === 'ios' ? 'ios-pin' : 'md-pin'}
+    />
+  ),
+};
+
+
+// New Task (UploadPhoto)
+const NewTaskStack = createStackNavigator({
+  // upload photo first
+  UploadPhoto:{
     screen: UploadPhoto,
-    // path: "/detail"
-  }
+  },
+  // then confirm or drop location
+  MediaGPS:{
+    screen: MediaGPS,
+  }, 
+  // fill out the form 
+  CreateTaskScreen:{
+    screen: CreateTaskScreen,
+  } 
+  
 });
 
-HomeStack.navigationOptions = {
-  tabBarLabel: 'Home',
+NewTaskStack.navigationOptions = {
+  tabBarLabel: 'Create Task',
+  // tabBarVisible: false,
   tabBarIcon: ({ focused }) => (
     <TabBarIcon
       focused={focused}
-      name={
-        Platform.OS === 'ios'
-          ? `ios-information-circle${focused ? '' : '-outline'}`
-          : 'md-information-circle'
-      }
+      name={Platform.OS === 'ios' ? 'ios-add-circle' : 'md-add-circle'}
     />
   ),
 };
 
-const MyFavoritesStack = createStackNavigator({
-  UserProfileScreen: UserProfileScreen,
+// User Profile (UserProfileScreen)
+const UserStack = createStackNavigator({
+  UserProfileScreen:{
+    screen: UserProfileScreen,
+  } 
 });
 
-MyFavoritesStack.navigationOptions = {
-  tabBarLabel: 'UserProfileScreen',
+UserStack.navigationOptions = {
+  tabBarLabel: 'User',
   tabBarIcon: ({ focused }) => (
     <TabBarIcon
       focused={focused}
-      name={Platform.OS === 'ios' ? 'ios-link' : 'md-link'}
+      name={Platform.OS === 'ios' ? 'ios-person' : 'md-person'}
     />
   ),
 };
+
+
 
 
 export default createBottomTabNavigator({
-  HomeStack,
-  UserProfileScreen
+  FeedStack,
+  MapStack,
+  NewTaskStack,
+  // WatchStack,
+  UserStack
 });
