+++ screens/LoginScreen.js
@@ -5,6 +5,9 @@ import { TouchableHighlight } from 'react-native';
 import Header from '../components/Header';
 import API from '../utils/API';
 
+// import MainNavigator from './navigation/MainNavigator';
+import { NavigationActions } from "react-navigation";
+
 const reactStyles = require('../react_native_styles/styles');
 const styles = reactStyles.default;
 
@@ -43,7 +46,15 @@ export default class LoginScreen extends Component {
         if(response.status == 200) {
           console.log(response)
           // Navigate to user profile screen after logging in
-          this.props.navigation.navigate('UserProfileScreen');
+          this.props.navigation.navigate('TasksScreen');
+
+          const navigateAction = NavigationActions.navigate({
+            routeName: "Home",
+            // params: { data: userObj }
+          });
+          this.props.navigation.dispatch(navigateAction);
+
+
         }
         // else {
         //   //print status text somewhere so user can see that login failed
