+++ screens/MediaGPS.js
@@ -118,8 +118,8 @@ export default class MediaGPS extends Component {
           region={{ 
             latitude: this.state.location.coords.latitude, 
             longitude: this.state.location.coords.longitude, 
-            latitudeDelta: 0.002, 
-            longitudeDelta: 0.004 
+            latitudeDelta: 0.0922, 
+            longitudeDelta: 0.0421 
           }}
           onPress={this.handlePress}
         >
